Socket sessions and validation with cookies
I have added the ability to validate a user (securely) with cookies. Uses can now add themselves. The templates are very messed up at this moment in time (please do not change any of them as it will make it way more difficult to fix for me in the future). This was a big update. E.g. The version needs to be changed to major, if not minor.