Leaderboards added, leaderboards routes added, scores calcualte, loads and loads...
If they got the incorrect answer, it shows them the correct answer; so users can learn (this was specified in our design documentation).

You can play with multiple players, chose questions, get scored, tells you if its right or wrong, has a leaderboard that shows the top 3 players and your place in the game. Fixed bug that showed a users question as incorrect, when it was correct.

After playing a game, the server resets everything to be get ready to play again.